Separa getPositionData, getStateData e getEquipmentData
