+++ fleet-management-app/src/app/page.js
@@ -1,4 +1,3 @@
-import Image from "next/image";
 import Map from "@/components/Map";
 
 export default function Home() {
